LayoutTests/svg/animations/accumulate-values-width-animation-expected.txt

SVG 1.1 dynamic animation tests


This tests values animation and accumulate='sum'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.width.animVal.value is 0
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 30
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 20
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 20
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 50
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 40
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 40
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 70
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 60
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 60
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 90
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 80
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 80
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 110
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 20
PASS rect.width.animVal.value is 100
PASS rect.width.baseVal.value is 20
PASS successfullyParsed is true

TEST COMPLETE
